decoration: Sync decorations with window resizes
Warning: bunch of code dup
