+++ src/handlers/xdg_shell.rs
@@ -6,25 +6,15 @@ use smithay::{
     },
     input::{Seat, pointer::Focus},
     reexports::{
-        calloop::Interest,
         wayland_protocols::xdg::shell::server::xdg_toplevel::ResizeEdge,
-        wayland_server::{
-            Resource,
-            protocol::{wl_output::WlOutput, wl_seat::WlSeat},
-        },
+        wayland_server::protocol::{wl_output::WlOutput, wl_seat::WlSeat},
     },
-    utils::{HookId, Serial},
-    wayland::{
-        compositor::{
-            self, BufferAssignment, CompositorHandler, SurfaceAttributes, add_pre_commit_hook,
-        },
-        shell::xdg::{
-            PopupSurface, PositionerState, ToplevelSurface, XdgShellHandler, XdgShellState,
-            XdgToplevelSurfaceData,
-        },
+    utils::Serial,
+    wayland::shell::xdg::{
+        PopupSurface, PositionerState, ToplevelSurface, XdgShellHandler, XdgShellState,
     },
 };
-use tracing::{error, field::Empty, trace, trace_span, warn};
+use tracing::warn;
 
 use crate::{
     focus::keyboard::KeyboardFocusTarget,
@@ -412,153 +402,3 @@ impl XdgShellHandler for State {
     }
 }
 delegate_xdg_shell!(State);
-
-/// Adds a pre-commit hook for mapped toplevels that blocks windows when transactions are pending.
-///
-/// It also takes over the role of the default dmabuf pre-commit hook, so when adding this
-/// be sure to remove the default hook.
-//
-// Yoinked from niri
-pub fn add_mapped_toplevel_pre_commit_hook(toplevel: &ToplevelSurface) -> HookId {
-    // TODO: move into separate thing
-    #[cfg(feature = "snowcap")]
-    compositor::add_post_commit_hook::<State, _>(toplevel.wl_surface(), |state, _dh, surface| {
-        let Some(window) = state.pinnacle.window_for_surface(surface) else {
-            error!("pre-commit hook for mapped surfaces must be removed upon unmapping");
-            return;
-        };
-
-        let geometry = compositor::with_states(surface, |states| {
-            let mut guard = states
-                .cached_state
-                .get::<smithay::wayland::shell::xdg::SurfaceCachedState>();
-            guard.current().geometry
-        });
-
-        if let Some(geometry) = geometry {
-            let size = geometry.size;
-
-            window.with_state(|state| {
-                if let Some(surface) = state.decoration_surface.as_ref() {
-                    surface.decoration_surface().with_pending_state(|state| {
-                        state.toplevel_size = Some(size);
-                    });
-                    surface.decoration_surface().send_pending_configure();
-                }
-            });
-        }
-    });
-
-    add_pre_commit_hook::<State, _>(toplevel.wl_surface(), move |state, _dh, surface| {
-        let _span = tracy_client::span!("mapped toplevel pre-commit");
-        let span =
-            trace_span!("toplevel pre-commit", surface = %surface.id(), serial = Empty).entered();
-
-        let Some(window) = state.pinnacle.window_for_surface(surface) else {
-            error!("pre-commit hook for mapped surfaces must be removed upon unmapping");
-            return;
-        };
-
-        let (got_unmapped, dmabuf, commit_serial) = compositor::with_states(surface, |states| {
-            let (got_unmapped, dmabuf) = {
-                let mut guard = states.cached_state.get::<SurfaceAttributes>();
-                match guard.pending().buffer.as_ref() {
-                    Some(BufferAssignment::NewBuffer(buffer)) => {
-                        let dmabuf = smithay::wayland::dmabuf::get_dmabuf(buffer).cloned().ok();
-                        (false, dmabuf)
-                    }
-                    Some(BufferAssignment::Removed) => (true, None),
-                    None => (false, None),
-                }
-            };
-
-            let role = states
-                .data_map
-                .get::<XdgToplevelSurfaceData>()
-                .unwrap()
-                .lock()
-                .unwrap();
-
-            (got_unmapped, dmabuf, role.configure_serial)
-        });
-
-        let mut transaction_for_dmabuf = None;
-        if let Some(serial) = commit_serial {
-            if !span.is_disabled() {
-                span.record("serial", format!("{serial:?}"));
-            }
-
-            trace!("taking pending transaction");
-            if let Some(transaction) = window.take_pending_transaction(serial) {
-                // Transaction can be already completed if it ran past the deadline.
-                if !transaction.is_completed() {
-                    let is_last = transaction.is_last();
-
-                    // If this is the last transaction, we don't need to add a separate
-                    // notification, because the transaction will complete in our dmabuf blocker
-                    // callback, which already calls blocker_cleared(), or by the end of this
-                    // function, in which case there would be no blocker in the first place.
-                    if !is_last {
-                        // Waiting for some other surface; register a notification and add a
-                        // transaction blocker.
-
-                        if let Some(client) = surface.client() {
-                            transaction.add_notification(
-                                state.pinnacle.blocker_cleared_tx.clone(),
-                                client.clone(),
-                            );
-                            compositor::add_blocker(surface, transaction.blocker());
-                        }
-                    }
-
-                    // Delay dropping (and completing) the transaction until the dmabuf is ready.
-                    // If there's no dmabuf, this will be dropped by the end of this pre-commit
-                    // hook.
-                    transaction_for_dmabuf = Some(transaction);
-                }
-            }
-        } else {
-            error!("commit on a mapped surface without a configured serial");
-        };
-
-        if let Some((blocker, source)) =
-            dmabuf.and_then(|dmabuf| dmabuf.generate_blocker(Interest::READ).ok())
-            && let Some(client) = surface.client()
-        {
-            let res = state
-                .pinnacle
-                .loop_handle
-                .insert_source(source, move |_, _, state| {
-                    // This surface is now ready for the transaction.
-                    drop(transaction_for_dmabuf.take());
-
-                    let display_handle = state.pinnacle.display_handle.clone();
-                    state
-                        .client_compositor_state(&client)
-                        .blocker_cleared(state, &display_handle);
-
-                    Ok(())
-                });
-            if res.is_ok() {
-                compositor::add_blocker(surface, blocker);
-                trace!("added dmabuf blocker");
-            }
-        }
-
-        if got_unmapped {
-            let Some(output) = window.output(&state.pinnacle) else {
-                return;
-            };
-
-            state.backend.with_renderer(|renderer| {
-                window.capture_snapshot_and_store(
-                    renderer,
-                    output.current_scale().fractional_scale().into(),
-                    1.0,
-                );
-            });
-        } else {
-            window.with_state_mut(|state| state.snapshot.take());
-        }
-    })
-}
