+++ src/lib.rs
@@ -8,6 +8,7 @@ pub mod decoration;
 pub mod focus;
 pub mod grab;
 pub mod handlers;
+pub mod hook;
 pub mod input;
 pub mod layout;
 pub mod output;
